Changed the way custom fields are handled Any field that is not a standard GELF field, nor is @version or @timestamp, will be made a custom field in the GELF message.
The rationale for this is that a user should not have to configure specific fields to pass in configuraton.
If the user has fields that they do not wish to send, they should use logstash-filter-mutate to do the removal in the pipeline.

This also makes the code a tad simpler, and means that the codec should be able to drop in with any output plugin.